refactor: migrate color format, tailwind version, and package manager to preferences store
