+++ src/store/editor-store.ts
@@ -4,19 +4,10 @@ import { ThemeEditorState, EditorType } from "@/types/editor";
 import { isEqual } from "@ngard/tiny-isequal";
 import { defaultThemeState } from "@/config/theme";
 import { getPresetThemeStyles } from "../utils/theme-presets";
-import { ColorFormat } from "@/types";
-
-type PackageManager = "pnpm" | "npm" | "yarn" | "bun";
 
 interface EditorStore {
   themeState: ThemeEditorState;
-  tailwindVersion: "3" | "4";
-  colorFormat: ColorFormat;
-  packageManager: PackageManager;
   setThemeState: (state: ThemeEditorState) => void;
-  setTailwindVersion: (version: "3" | "4") => void;
-  setColorFormat: (format: ColorFormat) => void;
-  setPackageManager: (pm: PackageManager) => void;
   applyThemePreset: (preset: string) => void;
   resetToDefault: () => void;
   resetToCurrentPreset: () => void;
@@ -28,20 +19,8 @@ export const useEditorStore = create<EditorStore>()(
   persist(
     (set, get) => ({
       themeState: defaultThemeState,
-      tailwindVersion: "4",
-      colorFormat: "oklch",
-      packageManager: "pnpm",
       setThemeState: (state: ThemeEditorState) => {
         set({ themeState: state });
-      },
-      setTailwindVersion: (version: "3" | "4") => {
-        set({ tailwindVersion: version });
-      },
-      setColorFormat: (format: ColorFormat) => {
-        set({ colorFormat: format });
-      },
-      setPackageManager: (pm: PackageManager) => {
-        set({ packageManager: pm });
       },
       applyThemePreset: (preset: string) => {
         const themeState = get().themeState;
